third_party/WebKit/LayoutTests/fast/dom/Window/window-postmessage-clone-really-deep-array-expected.txt

Tests that we abort cloning overdeep arrays.

PASS: 'postMessage(reallyDeepArray)' threw DataCloneError: Failed to execute 'postMessage' on 'Window': Value being cloned is either cyclic or too deeply nested.
PASS: eventData is done of type string
